+++ app/models/pay/charge.rb
@@ -74,6 +74,14 @@ module Pay
       refunded? && !full_refund?
     end
 
+    def amount_with_currency
+      Pay::Currency.format(amount, currency: currency)
+    end
+
+    def amount_refunded_with_currency
+      Pay::Currency.format(amount_refunded, currency: currency)
+    end
+
     def charged_to
       case payment_method_type
       when "card"