+++ app/models/pay/merchant.rb
@@ -10,12 +10,19 @@ module Pay
 
     delegate_missing_to :pay_processor
 
-    def self.processor_for(name)
+    def self.pay_processor_for(name)
       "Pay::#{name.to_s.classify}::Merchant".constantize
     end
 
     def pay_processor
-      @pay_processor ||= self.class.processor_for(processor).new(self)
+      return if processor.blank?
+      @pay_processor ||= self.class.pay_processor_for(processor).new(self)
+    end
+
+    def onboarding_complete?
+      ActiveModel::Type::Boolean.new.cast(
+        (data.presence || {})["onboarding_complete"]
+      )
     end
   end
 end