+++ lib/pay/braintree/webhooks/subscription_charged_unsuccessfully.rb
@@ -11,11 +11,11 @@ module Pay
           pay_subscription = Pay::Subscription.find_by_processor_and_id(:braintree, subscription.id)
           return unless pay_subscription.present?
 
-          # billable = pay_subscription.owner
-          # charge = Pay::Braintree::Billable.new(billable).save_transaction(subscription.transactions.first)
+          # pay_customer = pay_subscription.customer
+          # pay_charge = Pay::Braintree::Billable.new(pay_customer).save_transaction(subscription.transactions.first)
 
           # if Pay.send_emails
-          #  Pay::UserMailer.with(billable: billable, charge: charge).receipt.deliver_later
+          #  Pay::UserMailer.with(pay_customer: pay_charge.customer, charge: pay_charge).receipt.deliver_later
           # end
         end
       end