+++ test/dummy/app/controllers/paddle/charges_controller.rb
@@ -2,7 +2,7 @@ class Paddle::ChargesController < ApplicationController
   before_action :set_charge, only: [:show, :refund]
 
   def index
-    @charges = Pay::Charge.where(processor: :paddle).order(created_at: :desc)
+    @charges = Pay::Charge.joins(:customer).where(pay_customers: {processor: :paddle}).order(created_at: :desc)
   end
 
   def show
@@ -12,9 +12,9 @@ class Paddle::ChargesController < ApplicationController
   end
 
   def create
-    current_user.processor = params[:processor]
-    current_user.card_token = params[:card_token]
-    charge = current_user.charge(params[:amount])
+    current_user.set_payment_processor params[:processor]
+    current_user.payment_processor.payment_method_token = params[:card_token]
+    charge = current_user.payment_processor.charge(params[:amount])
     redirect_to paddle_charge_path(charge)
   rescue Pay::Error => e
     flash[:alert] = e.message