+++ test/dummy/app/controllers/stripe/subscriptions_controller.rb
@@ -3,7 +3,7 @@ module Stripe
     before_action :set_subscription, only: [:show, :edit, :update, :destroy, :cancel, :resume]
 
     def index
-      @subscriptions = Pay::Subscription.where(processor: :stripe).order(created_at: :desc)
+      @subscriptions = Pay::Subscription.joins(:customer).where(pay_customers: {processor: :stripe}).order(created_at: :desc)
     end
 
     def show
@@ -13,9 +13,9 @@ module Stripe
     end
 
     def create
-      current_user.processor = params[:processor]
-      current_user.card_token = params[:card_token]
-      subscription = current_user.subscribe(plan: params[:price_id])
+      current_user.set_payment_processor params[:processor]
+      current_user.payment_processor.payment_method_token = params[:card_token]
+      subscription = current_user.payment_processor.subscribe(plan: params[:price_id])
       redirect_to stripe_subscription_path(subscription)
     rescue Pay::ActionRequired => e
       redirect_to pay.payment_path(e.payment.id)
@@ -48,7 +48,7 @@ module Stripe
     private
 
     def set_subscription
-      @subscription = Pay::Subscription.where(processor: :stripe).find(params[:id])
+      @subscription = Pay::Subscription.find(params[:id])
     end
   end
 end