+++ test/dummy/app/models/team.rb
@@ -2,7 +2,7 @@ class Team < ApplicationRecord
   pay_customer
   pay_merchant
 
-  belongs_to :owner, class_name: "User"
+  belongs_to :owner, polymorphic: true
 
   def email
     owner.email