+++ test/mailers/pay/user_mailer_test.rb
@@ -3,12 +3,13 @@ require "test_helper"
 class UserMailerTest < ActionMailer::TestCase
   setup do
     @charge = pay_charges(:stripe)
-    @user = @charge.customer.owner
+    @pay_customer = @charge.customer
+    @user = @pay_customer.owner
     @user.update(extra_billing_info: "extra billing info")
   end
 
   test "receipt" do
-    email = Pay::UserMailer.with(billable: @user, charge: @charge).receipt
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, charge: @charge).receipt
 
     assert_equal [@user.email], email.to
     assert_equal I18n.t("pay.user_mailer.receipt.subject"), email.subject
@@ -24,13 +25,13 @@ class UserMailerTest < ActionMailer::TestCase
     @charge.stubs(:filename).returns(filename)
     @charge.stubs(:receipt).returns(receipt)
 
-    email = Pay::UserMailer.with(billable: @user, charge: @charge).receipt
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, charge: @charge).receipt
 
     assert_equal filename, email.attachments.first.filename
   end
 
   test "refund" do
-    email = Pay::UserMailer.with(billable: @user, charge: @charge).refund
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, charge: @charge).refund
 
     assert_equal [@user.email], email.to
     assert_equal I18n.t("pay.user_mailer.refund.subject"), email.subject
@@ -38,7 +39,7 @@ class UserMailerTest < ActionMailer::TestCase
 
   test "subscription_renewing" do
     time = Time.current
-    email = Pay::UserMailer.with(billable: @user, subscription: Pay::Subscription.new, date: time).subscription_renewing
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, subscription: Pay::Subscription.new, date: time).subscription_renewing
 
     assert_equal [@user.email], email.to
     assert_equal I18n.t("pay.user_mailer.subscription_renewing.subject"), email.subject
@@ -46,10 +47,28 @@ class UserMailerTest < ActionMailer::TestCase
   end
 
   test "payment_action_required" do
-    email = Pay::UserMailer.with(billable: @user, payment_intent_id: "x", subscription: Pay::Subscription.new).payment_action_required
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, payment_intent_id: "x", subscription: Pay::Subscription.new).payment_action_required
 
     assert_equal [@user.email], email.to
     assert_equal I18n.t("pay.user_mailer.payment_action_required.subject"), email.subject
     assert_includes email.body.decoded, Pay::Engine.instance.routes.url_helpers.payment_path("x")
+  end
+
+  test "receipt with no extra billing info column" do
+    team = teams(:one)
+    @pay_customer.update!(owner: team)
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, charge: @charge).receipt
+
+    assert_equal [team.owner.email], email.to
+    assert_equal I18n.t("pay.user_mailer.receipt.subject"), email.subject
+  end
+
+  test "refund with no extra billing info column" do
+    team = teams(:one)
+    @pay_customer.update!(owner: team)
+    email = Pay::UserMailer.with(pay_customer: @pay_customer, charge: @charge).refund
+
+    assert_equal [team.owner.email], email.to
+    assert_equal I18n.t("pay.user_mailer.refund.subject"), email.subject
   end
 end