+++ test/models/pay/merchant_test.rb
@@ -1,6 +1,14 @@
 require "test_helper"
 
 class Pay::MerchantTest < ActiveSupport::TestCase
+  test "should return stripe connect onboarding status" do
+    merchant = Pay::Merchant.new
+    refute merchant.onboarding_complete?
+
+    merchant.onboarding_complete = true
+    assert merchant.onboarding_complete?
+  end
+  
   test "text data column" do
     pay_merchants(:one).update!(onboarding_complete: true)
     assert_equal true, pay_merchants(:one).onboarding_complete