+++ test/pay/paddle/subscription_test.rb
@@ -8,14 +8,18 @@ class Pay::Paddle::Subscription::Test < ActiveSupport::TestCase
   test "paddle cancel" do
     subscription = @pay_customer.subscription
     next_payment_date = Time.zone.parse(subscription.processor_subscription.next_payment[:date])
-    subscription.cancel
+    assert_difference "@pay_customer.payment_methods.count", -1 do
+      subscription.cancel
+    end
     assert_equal subscription.ends_at, next_payment_date
     assert_equal "canceled", subscription.status
   end
 
   test "paddle cancel_now!" do
     subscription = @pay_customer.subscription
-    subscription.cancel_now!
+    assert_difference "@pay_customer.payment_methods.count", -1 do
+      subscription.cancel_now!
+    end
     assert subscription.ends_at <= Time.current
     assert_equal "canceled", subscription.status
   end
@@ -67,5 +71,22 @@ class Pay::Paddle::Subscription::Test < ActiveSupport::TestCase
     @pay_customer.subscription.swap("594470")
     assert_equal 594470, @pay_customer.subscription.processor_subscription.plan_id
     assert_equal "active", @pay_customer.subscription.status
+  end
+
+  test "paused from timestamp" do
+    pay_subscriptions(:paddle).update(paddle_paused_from: 14.days.from_now)
+    assert_equal ActiveSupport::TimeWithZone, pay_subscriptions(:paddle).paddle_paused_from.class
+  end
+
+  test "paddle cancel paused subscription" do
+    subscription = @pay_customer.subscription
+    subscription.update(paddle_paused_from: "Sat, 04 Jul 2020 00:00:00.000000000 UTC +00:00")
+    assert_nil subscription.processor_subscription.next_payment
+    assert_difference "@pay_customer.payment_methods.count", -1 do
+      subscription.cancel
+    end
+    assert_not_nil subscription.ends_at
+    assert_equal subscription.ends_at, subscription.paddle_paused_from
+    assert_equal "canceled", subscription.status
   end
 end